pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
description="flask template"
[tool.poetry.dependencies]
flask = "==1.0.2"
python = "^3.8"
python-dotenv = "^0.17.0"
werkzeug = "^1.0.1"
wtforms = "^2.3.3"
flask-login = "^0.5.0"
flask-wtf = "^0.14.3"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
flask-moment = "^0.11.0"
flask-bootstrap = "^3.3.7"
flask-migrate = "^2.7.0"
flask-sqlalchemy = "^2.5.1"
email-validator = "^1.1.2"
sqlalchemy = "^1.4.11"
alembic = "^1.5.8"
flask-debugtoolbar = "^0.11.0"
xlrd = "^2.0.1"
